Major revamp
The following changes were made after the application was none responsive.

A few items were not part of the original template and were created manually
but not imported including the load balancer and target group.

The auto scaling group was included in the template but it pointed directly
at the EC2 instance, not the target group. It is now configured correctly.

The Route53 hosted zone was removed from the template so it's not
constanly renewed during development. Keeping it in the template
would require updating the DNS name servers in Namecheap and would also
mean the DNS records would updated each time they are removed. This would
require about 30 to 45 minutes for the DNS records to be updated before
being able to access the domain.

After applying the template, the only manual step that is required to is to add
and A (alias) record in Route53 that points the domain to the load balancer.

The ACM certificate and the Route53 hosted zone and records are not managed by
this template.

Main.tf
Delete EC2 instance, Elastice IP, Route53 hosted zone, route53 A record.

Add second subnet and routing table association, security group rules
for HTTP and HTTPS.

Add launch template, auto scaling group, target group, application
load balancer, load balancer listeners for HTTPS and HTTP redirect.

outputs.tf
Delete ELASTIC_IP. Edit keyname so it gets its value from the
launch template, instead of the EC2 instance.

variables.tf
Edit the AMI id.