Add "A" record
The hosted zone, NS and NOA records are not included because
it takes about 45 minutes to 1 hour before they are available
in DNS. For this reason, it's better to management manually
without Terraform.

The "A" record points to the loadbalancer and works well with
Terraform. It does however require about a 5 minute wait for
the site to become available through DNS.